testes fodas.exe

+++ src/estruturas.rs
@@ -3,6 +3,6 @@ use serde::{Serialize, Deserialize};
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Tecla {
     pub pressed: bool,
-    pub key: u16,
+    pub key: u8,
     pub special: bool
 }
